+++ cli/cli/commands/exec/exec.go
@@ -22,7 +22,7 @@ import (
 
 const (
 	scriptOrModulePathKey = "script-or-module-path"
-	startosisExtension    = ".star"
+	starlarkExtension     = ".star"
 
 	moduleArgsFlagKey = "args"
 	defaultModuleArgs = "{}"
@@ -50,10 +50,10 @@ var (
 	kurtosisNoExecutionError      *kurtosis_core_rpc_api_bindings.KurtosisExecutionError
 )
 
-var StartosisExecCmd = &lowlevel.LowlevelKurtosisCommand{
-	CommandStr:       command_str_consts.StartosisExecCmdStr,
-	ShortDescription: "Execute a Startosis script or module",
-	LongDescription: "Execute a Startosis module or script in an enclave. For a script we expect a path to a " + startosisExtension +
+var StarlarkExecCmd = &lowlevel.LowlevelKurtosisCommand{
+	CommandStr:       command_str_consts.StarlarkExecCmdStr,
+	ShortDescription: "Execute a Starlark script or module",
+	LongDescription: "Execute a Starlark module or script in an enclave. For a script we expect a path to a " + starlarkExtension +
 		" file. For a module we expect path to a directory containing kurtosis.mod. If the enclave-id param is provided, Kurtosis " +
 		"will exec the script inside this enclave, or create it if it doesn't exist. If no enclave-id param is " +
 		"provided, Kurtosis will create a new enclave with a default name derived from the script or module name.",
@@ -131,9 +131,9 @@ func run(
 		return stacktrace.Propagate(err, "An error occurred getting the is-partitioning-enabled setting using flag key '%v'", isPartitioningEnabledFlagKey)
 	}
 
-	startosisScriptOrModulePath, err := args.GetNonGreedyArg(scriptOrModulePathKey)
+	starlarkScriptOrModulePath, err := args.GetNonGreedyArg(scriptOrModulePathKey)
 	if err != nil {
-		return stacktrace.Propagate(err, "Error reading the Startosis script or module dir at '%s'. Does it exist?", startosisScriptOrModulePath)
+		return stacktrace.Propagate(err, "Error reading the Starlark script or module dir at '%s'. Does it exist?", starlarkScriptOrModulePath)
 	}
 
 	dryRun, err := flags.GetBool(dryRunFlagKey)
@@ -158,25 +158,25 @@ func run(
 		defer output_printers.PrintEnclaveId(enclaveCtx.GetEnclaveID())
 	}
 
-	fileOrDir, err := os.Stat(startosisScriptOrModulePath)
+	fileOrDir, err := os.Stat(starlarkScriptOrModulePath)
 	if err != nil {
-		return stacktrace.Propagate(err, "There was an error reading file or module from disk at '%v'", startosisScriptOrModulePath)
+		return stacktrace.Propagate(err, "There was an error reading file or module from disk at '%v'", starlarkScriptOrModulePath)
 	}
 
 	if fileOrDir.Mode().IsRegular() {
-		if !strings.HasSuffix(startosisScriptOrModulePath, startosisExtension) {
-			return stacktrace.NewError("Expected a script with a '%s' extension but got file '%v' with a different extension", startosisExtension, startosisScriptOrModulePath)
+		if !strings.HasSuffix(starlarkScriptOrModulePath, starlarkExtension) {
+			return stacktrace.NewError("Expected a script with a '%s' extension but got file '%v' with a different extension", starlarkExtension, starlarkScriptOrModulePath)
 		}
-		err = executeScript(enclaveCtx, startosisScriptOrModulePath, dryRun)
+		err = executeScript(enclaveCtx, starlarkScriptOrModulePath, dryRun)
 		if err != nil {
-			return stacktrace.Propagate(err, "An error occurred while executing script '%v'", startosisScriptOrModulePath)
+			return stacktrace.Propagate(err, "An error occurred while executing script '%v'", starlarkScriptOrModulePath)
 		}
 		return nil
 	}
 
-	err = executeModule(enclaveCtx, startosisScriptOrModulePath, serializedJsonArgs, dryRun)
+	err = executeModule(enclaveCtx, starlarkScriptOrModulePath, serializedJsonArgs, dryRun)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred while running the module '%v'", startosisScriptOrModulePath)
+		return stacktrace.Propagate(err, "An error occurred while running the module '%v'", starlarkScriptOrModulePath)
 	}
 
 	return nil
@@ -211,12 +211,12 @@ func validateScriptOrModulePath(_ context.Context, _ *flags.ParsedFlags, args *a
 func executeScript(enclaveCtx *enclaves.EnclaveContext, scriptPath string, dryRun bool) error {
 	fileContentBytes, err := os.ReadFile(scriptPath)
 	if err != nil {
-		return stacktrace.Propagate(err, "Unable to read content of Startosis script file '%s'", scriptPath)
+		return stacktrace.Propagate(err, "Unable to read content of Starlark script file '%s'", scriptPath)
 	}
 
 	executionResponse, err := enclaveCtx.ExecuteStartosisScript(string(fileContentBytes), dryRun)
 	if err != nil {
-		return stacktrace.Propagate(err, "An unexpected error occurred executing the Startosis script '%s'", scriptPath)
+		return stacktrace.Propagate(err, "An unexpected error occurred executing the Starlark script '%s'", scriptPath)
 	}
 
 	err = validateExecutionResponse(executionResponse, scriptPath, scriptArgForLogging, dryRun)
@@ -230,7 +230,7 @@ func executeScript(enclaveCtx *enclaves.EnclaveContext, scriptPath string, dryRu
 func executeModule(enclaveCtx *enclaves.EnclaveContext, modulePath string, serializedParams string, dryRun bool) error {
 	executionResponse, err := enclaveCtx.ExecuteStartosisModule(modulePath, serializedParams, dryRun)
 	if err != nil {
-		return stacktrace.Propagate(err, "An unexpected error occurred executing the Startosis module '%s'", modulePath)
+		return stacktrace.Propagate(err, "An unexpected error occurred executing the Starlark module '%s'", modulePath)
 	}
 
 	err = validateExecutionResponse(executionResponse, modulePath, moduleArgForLogging, dryRun)
@@ -243,10 +243,10 @@ func executeModule(enclaveCtx *enclaves.EnclaveContext, modulePath string, seria
 
 func validateExecutionResponse(executionResponse *kurtosis_core_rpc_api_bindings.ExecuteStartosisResponse, scriptOrModulePath string, scriptOrModuleArg string, dryRun bool) error {
 	if executionResponse.GetInterpretationError() != kurtosisNoInterpretationError {
-		return stacktrace.NewError("There was an error interpreting the Startosis %s '%s': \n%v", scriptOrModuleArg, scriptOrModulePath, executionResponse.GetInterpretationError().GetErrorMessage())
+		return stacktrace.NewError("There was an error interpreting the Starlark %s '%s': \n%v", scriptOrModuleArg, scriptOrModulePath, executionResponse.GetInterpretationError().GetErrorMessage())
 	}
 	if executionResponse.GetValidationErrors() != kurtosisNoValidationError {
-		return stacktrace.NewError("There was an error validating the Startosis %s '%s': \n%v", scriptOrModuleArg, scriptOrModulePath, executionResponse.GetValidationErrors().GetErrors())
+		return stacktrace.NewError("There was an error validating the Starlark %s '%s': \n%v", scriptOrModuleArg, scriptOrModulePath, executionResponse.GetValidationErrors().GetErrors())
 	}
 
 	var scriptOutputLines []string
@@ -260,7 +260,7 @@ func validateExecutionResponse(executionResponse *kurtosis_core_rpc_api_bindings
 	logrus.Infof("Kurtosis script successfully interpreted and validated. List of Kurtosis instructions generated:\n%v", strings.Join(concatenatedKurtosisInstructions, "\n"))
 
 	if executionResponse.GetExecutionError() != kurtosisNoExecutionError {
-		return stacktrace.NewError("There was an error executing the Startosis %s '%s': \n%v", scriptOrModuleArg, scriptOrModulePath, executionResponse.GetExecutionError().GetErrorMessage())
+		return stacktrace.NewError("There was an error executing the Starlark %s '%s': \n%v", scriptOrModuleArg, scriptOrModulePath, executionResponse.GetExecutionError().GetErrorMessage())
 	}
 
 	if dryRun {
@@ -290,7 +290,7 @@ func getOrCreateEnclaveContext(
 		}
 		return enclaveContext, false, nil
 	}
-	logrus.Infof("Creating a new enclave for the startosis script to execute inside...")
+	logrus.Infof("Creating a new enclave for the Starlark script to execute inside...")
 	enclaveContext, err := kurtosisContext.CreateEnclave(ctx, enclaveId, isPartitioningEnabled)
 	if err != nil {
 		return nil, false, stacktrace.Propagate(err, fmt.Sprintf("Unable to create new enclave with ID '%s'", enclaveId))